+++ app/smartattendance/serializers/UsuarioSerializer.py
@@ -1,6 +1,6 @@
 from rest_framework.serializers import ModelSerializer
 
-from .models import Usuario
+from ..models import Usuario
 
 class UsuarioSerializer(ModelSerializer):
     class Meta:
